+++ matrix.h
@@ -25,12 +25,12 @@ matrix_t* matrix_create_from_file(char* name);
 
 void print_matrix(matrix_t* matrix);
 
-void matrix_set(matrix_t* matrix, int i, int j, float v);
-float matrix_get(matrix_t* matrix, int i, int j);
+void matrix_set(matrix_t* matrix, int i, int j, double v);
+double matrix_get(matrix_t* matrix, int i, int j);
 
 // Operations
 matrix_t* matrix_add(matrix_t* A, matrix_t* B);
 matrix_t* matrix_subtract(matrix_t* A, matrix_t* B);
-matrix_t* matrix_mult_scalar(float n, matrix_t* A);
+matrix_t* matrix_mult_scalar(double n, matrix_t* A);
 
 #endif
